# Add spectrum to beam object