--- a/model/experiment_list.py
+++ b/model/experiment_list.py
@@ -372,23 +372,30 @@ class ExperimentListDict(object):
             )
 
         # Read the beam spectra if available
-        self._read_spectra_pickle(el.beams())
+        self._read_spectra_h5(el.beams())
 
         # Return the experiment list
         return el
 
-    def _read_spectra_pickle(self, beams):
+    def _read_spectra_h5(self, beams):
         """ Read the beam spectra if available """
-        if "spectra_pickle" not in self._obj:
+        if "spectra_h5" not in self._obj:
             return
-        with open(self._obj["spectra_pickle"], "rb") as infile:
-            all_energies, all_weights = pickle.load(infile)
-        assert len(all_energies) == 1 or len(all_energies) == len(all_weights)
+
+        import h5py
+
+        f = h5py.File(self._obj["spectra_h5"], "r")
+        all_energies = f["all_energies"][()]
+        all_weights = f["all_weights"][()]
+
+        assert (
+            len(all_energies.shape) == 1 or all_energies.shape == all_weights.shape
+        ), (all_energies.shape, all_weights.shape)
         for i, beam in enumerate(beams):
             spectrum_index = self._obj["beam"][i].get("spectrum_index")
             if spectrum_index is not None:
-                if len(all_energies) == 1:
-                    beam.set_spectrum(all_energies[0], all_weights[spectrum_index])
+                if len(all_energies.shape) == 1:
+                    beam.set_spectrum(all_energies, all_weights[spectrum_index])
                 else:
                     beam.set_spectrum(
                         all_energies[spectrum_index], all_weights[spectrum_index]