@phyy-nx 2020-04-08
This is an alternate version of #157 that doesn't use class derivation.  It still does serialize the spectrum to the json file which increases file sizes dramatically.  This is a real problem, and some solutions are being discussed.