@graeme-winter 2020-04-15
> This is an alternate version of #157 that doesn't use class derivation. It still does serialize the spectrum to the json file which increases file sizes dramatically. This is a real problem, and some solutions are being discussed.

You have a proposal of how to fix this before merging the PR?