@phyy-nx 2020-04-15
TL;DR: we discussed several ways of fixing the serialization issue but nothing seems to be straightforward.  The timings/file sizes are impacted but processing seems fast enough for us that we want to punt the issue for now.

Details: this implementation saves two big arrays of numbers for each shot.  The .expt file sizes get a lot bigger.  I ran some tests using SwissFEL master files for the JF16M (1400+ images).  Some numbers:

No spectra:
- master h5 file: 1.2MB
- imported.expt: 2.0MB
- import time: 18.7s
- Read time*: 1.4s

With spectra:
- master h5 file: 58MB
- imported.expt: 144MB
- import time: 41.8s
- Read time*: 12.9s

(* time to de-serialize the 1400+ experiment list)

I do think I can save some space in the master h5 file by only storing the energy spectrum axis once, but that won't affect the imported.expt file size or the timings.

Unfortunately I can't come up with a way to fix this.  I'd prefer not to serialize the spectra at all, presuming it to be immutable, but then the beam object would need a link back to the raw data, which is not currently a thing in dxtbx.  I mucked around with this a bit, creating a custom NeXusBeam object in nexus.py that had a link back to the raw data and that link was serialized instead, but I gave up on it pretty quickly.