@graeme-winter 2020-04-16
Obviously the null value could be a standard monochromatic beam