@graeme-winter 2020-04-16
For what it’s worth this will also extend conceptually for the beam shape as well, where that could be a simple model or a complex measured image. So I think revisiting the fundamental design here could help

I welcome alternative perspectives