@phyy-nx 2020-04-16
@graeme-winter I'm interested in this as an option.  I do think we can punt it down the road a bit.  It is very easy to write backwards compatibility into the deserialization methods if the object model changes.