@phyy-nx 2020-04-16
I have created a new branch, spectrum_beam2_usepickles, which greatly saves on file space and read time.  See commit 47852f2.  If agreeable, I'll add that commit to this pull request.  The change saves the spectra to an external pickle file instead of the json ascii file.  This is similar to the lookup.mask feature that already exists. (Note it does exacerbate dials/dials#1238.)

New stats:
With external spectra pickle:

master h5 file: 30MB <-- savings comes from now only storing 1 energy array
imported.expt: 2MB, + 18M spectrum.pickle
import time: 31.4s
Read time*: 2.1s

These are much better numbers.  Note, I couldn't use the external lookup code directly as that applies only to image sets.