@graeme-winter 2020-04-16
> @graeme-winter I'm interested in this as an option. I do think we can punt it down the road a bit. It is very easy to write backwards compatibility into the deserialization methods if the object model changes.

Hows about make this an external object, which does it's own lookup via pickle as above - to build in the idea that it may evolve. If you embed it directly in beam it can never be removed... I have bad experiences here. 

NXspectrum class or similar which implements spectrum, with a spectrum pointer on beam?