@phyy-nx 2020-04-16
Thinking more about this, I'm not totally on board with a separate object.  I'm arriving at the conclusion that NeXus and dxtbx should be more tightly linked in our development efforts.  In other words, if there is something missing that we need to describe in our experiments, we should consider the NeXus solution that will be used to represent the metadata, and match the dxtbx interfaces to that spec.  If we need something in NeXus that isn't there to support a specific dxtbx interface that we need, then we should work on adding it to NeXus.

In this case, the provided examples are monchromatic source, source with 1D spectra, source with a Gaussian spectrum, source with a known beam profile, pink beam, and neutron (of note, there is a [TOF neutron NeXus application definition)](https://manual.nexusformat.org/classes/applications/NXtofnpd.html).  I think all these cases are [supported in NeXus already](https://manual.nexusformat.org/classes/applications/NXmx.html) (see for example the profile and incident_beam_size fields for the beam profile case, and incident_wavelength_spread for a gaussian spectra).  These things are all properties on the NXbeam class, so preserving that interface in dxtbx makes sense to me. After all, a spectrum and a beam shape are different, and they are both something that a beam has.

If you still think that a beam object should produce a spectrum object, then what would the base class of the spectrum object be, such that other things could derive from it, such as a beam profile?  Ideas:
- BeamShape
- BeamDetails
- BeamParameters
- BeamBucket
- BeamExtras
- BeamSpec