@alyubimov 2020-04-17
I've got a very naive question, from a chemist's perspective, about this:

> No spectra:
> 
> * master h5 file: 1.2MB
> * imported.expt: 2.0MB
> * import time: 18.7s
> * Read time*: 1.4s
> 
> With spectra:
> 
> * master h5 file: 58MB
> * imported.expt: 144MB
> * import time: 41.8s
> * Read time*: 12.9s
> 
> (* time to de-serialize the 1400+ experiment list)

What-all is recorded in the "spectrum" that it requires _this much_ memory? I would've thought a spectrum could be expressed as a 1-dimensional list: measurement start (e.g. 120 Angstrom), step size (e.g. 5 Angstrom), and then the list of counts. What else is there that requires 140+ MB?