@phyy-nx 2020-04-17
So it's a great question.  These spectra are 2500 energy channels and 2500 'weights' per image.  Here's an example spectra from SwissFEL (weights vs. eV):

![image](https://user-images.githubusercontent.com/13471434/79614456-70e96000-80b5-11ea-957e-f431d41e01ee.png)

In the imported json file, each number is represented in ASCII and seperated by things like newlines, spaces, and commas, that's what adds up to 144 MB.  The new version which uses an external pickle file for the spectra, is more like 18MB, plus it only saves the energy channels once.  Much more realistic for 1400\*2500*double precision for the whole set.